use binary search in coap_findOptions
It is OK to use a for loop through all element in option list, but
'FIXME' tag annoying.

Use binary search to get the lowerbound and upperbound, then the count
should be lowerbound - upperbound. If lowerbound point to an element
that not equal to what we want, return NULL because no any element matches.